Adds Deathmatch Demo and Supporting Features
* Adds Deathmatch demo project.
* Changes render pipeline to use 3D math to transform sprites. Only handles translations and there is no projection happening.
* Adds a camera that can follow a player controled character.
* Updates DirectInpuDebugDemo to use new features.
* Updates Unit tests to use new interfaces.
* Adds logic to stop jittering when moving diagonally. Locks the position to the line through the center of the pixel x = y (only works for square pixels. Would be better if slope was calculated from font width and height).

+++ Engine/Native/NativeWindows.cs
@@ -94,7 +94,7 @@ namespace Engine.Native
     {
         public byte ForegroundColorIndex;
         public byte BackgroundColorIndex;
-        public byte CharacterCode;
+        public ushort CharacterCode;
     }
     /// <summary>
     ///
